requirements.txt
africastalking==1.2.7
beautifulsoup4==4.12.2
Django==4.2.6
django-bootstrap-v5==1.0.11
Pillow==10.1.0
soupsieve==2.5
sqlparse==0.4.4

colorlog==6.8.2
cryptography==41.0.4
django-cors-headers==4.3.0
django-filter==23.3
djangorestframework==3.14.0
dj-database-url==2.1.0
gunicorn==21.2.0
Markdown==3.5
packaging==23.2
pycparser==2.21
PyMySQL==1.1.0
pytz==2023.3.post1
sqlparse==0.4.4

beautifulsoup4==4.12.2
cryptography==41.0.4
django-oauth-toolkit==2.3.0
icecream==2.1.3
ipython==8.16.1
lxml==4.9.3
mailerlite-api-python==0.10.0
oauthlib==3.2.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
requests==2.31.0
grequests==0.7.0

python-dotenv==1.0.0